+++ frontend/src/pages/ViewDietPlanPage.tsx
@@ -38,8 +38,8 @@ export const ViewDietPlanPage = () => {
 
   const onSuccess = () => {
     toast.success("תפריט נשמר בהצלחה!");
-    navigation(MainRoutes.USERS + `/${id}`);
     queryClient.invalidateQueries({ queryKey: [`${QueryKeys.USER_DIET_PLAN}${id}`] });
+    navigation(MainRoutes.USERS + `/${id}`);
   };
 
   const onError = (e: any) => {
@@ -91,7 +91,7 @@ export const ViewDietPlanPage = () => {
     queryFn: () => getAllDietPlanPresets().then((res) => res.data),
   });
 
-  const { isLoading, error } = useQuery({
+  const { isLoading, error, data } = useQuery({
     queryKey: [`${QueryKeys.USER_DIET_PLAN}${id}`],
     enabled: !!id,
     staleTime: FULL_DAY_STALE_TIME,
@@ -105,7 +105,8 @@ export const ViewDietPlanPage = () => {
           setIsNewPlan(true);
           setDietPlan(defaultDietPlan);
           return e;
-        }),
+        })
+        .finally(() => console.log(`ran`)),
   });
 
   const handleSelect = (presetName: string) => {
@@ -124,7 +125,7 @@ export const ViewDietPlanPage = () => {
 
   if (isLoading || dietPlanPresets.isLoading) return <Loader size="large" />;
   if (error) return <ErrorPage message={error.message} />;
-  const plan = dietPlan || defaultDietPlan;
+  const plan = dietPlan || data;
 
   return (
     <div className=" flex flex-col gap-4 size-full hide-scrollbar overflow-y-auto">
@@ -141,8 +142,8 @@ export const ViewDietPlanPage = () => {
           ))}
         </SelectContent>
       </Select>
-      <DietPlanForm dietPlan={plan} updateDietPlan={updateDietPlan} />
-      {plan.meals.length > 0 && (
+      {plan && <DietPlanForm dietPlan={plan} updateDietPlan={updateDietPlan} />}
+      {plan && plan.meals.length > 0 && (
         <div>
           <CustomButton
             className="font-bold"